honksquad: refactor: move QuickDialogCloseEvent to fork namespace
Extract QuickDialogCloseEvent from upstream QuickDialogOpenEvent.cs to
Content.Shared/RussStation/Economy/Events/, restoring the upstream file
to its original state. Add HONK-marked usings in client/server systems.

Closes #408
