+++ Content.Server/Administration/QuickDialogSystem.cs
@@ -1,5 +1,8 @@
 using System.Diagnostics.CodeAnalysis;
 using Content.Shared.Administration;
+//HONK START
+using Content.Shared.RussStation.Economy.Events;
+//HONK END
 using Robust.Server.Player;
 using Robust.Shared.Enums;
 using Robust.Shared.Network;
